add onlyif param to shift::entry.pp

+++ code/manifests/shift/entry.pp
@@ -1,13 +1,14 @@
 define lcgdm::shift::entry($component, $type) {
 
-    augeas { "shiftentry_${component}-${type}":
-      context => "/files/etc/shift.conf",
-      changes => [
-        "rm name[.='$component'][type='$type']",
-        "set name[last()+1] $component",
-        "set name[last()]/type $type",
+  augeas { "shiftentry_${component}-${type}":
+    context => "/files/etc/shift.conf",
+    changes => [
+      "rm name[.='$component'][type='$type']",
+      "set name[last()+1] $component",
+      "set name[last()]/type $type",
     ],
-      require => [ File["/usr/share/augeas/lenses/dist/shift.aug"], File["/etc/shift.conf"], ],
-    }
+    onlyif  => "match name[.='$component'][type='$type'] size == 0",
+    require => [ File["/usr/share/augeas/lenses/dist/shift.aug"], File["/etc/shift.conf"], ],
+  }
 
 }
